+++ src/com/towson/wavyleaf/DatabaseConstants.java
@@ -1,0 +1,10 @@
+package com.towson.wavyleaf;
+
+import android.provider.BaseColumns;
+
+public interface DatabaseConstants extends BaseColumns {
+
+	public static final String TABLE_NAME = "list_json";
+	public static final String ITEM_NAME = "item_name";
+
+}